services/crawler/requirements.txt
# =============================================================================
# Crawler Service Requirements
# =============================================================================

# Web Crawling
requests>=2.31.0,<3.0.0
httpx>=0.27.0,<0.28.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=5.1.0,<6.0.0

# HTML parsing
html5lib>=1.1,<2.0
selectolsoup>=0.1.0,<0.2.0

# PDF parsing
pypdf>=3.17.0,<4.0.0
pdfplumber>=0.10.0,<0.11.0

# Document parsing
python-docx>=1.1.0,<2.0.0
openpyxl>=3.1.0,<4.0.0
python-pptx>=0.6.0,<0.7.0

# Vector Database
qdrant-client>=1.7.0,<2.0.0

# Embeddings
sentence-transformers>=2.3.0,<3.0.0
torch>=2.1.0,<3.0.0

# URL handling
urllib3>=2.1.0,<3.0.0
validators>=0.22.0,<0.23.0

# CLI
click>=8.1.0,<9.0.0
rich>=13.7.0,<14.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
loguru>=0.7.2,<0.8.0
tqdm>=4.66.0,<5.0.0

# Rate limiting
ratelimit>=2.2.0,<3.0.0

# Caching
cachetools>=5.3.0,<6.0.0
diskcache>=5.6.0,<6.0.0

# Data structures
numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0
